feat: add access rights to students

+++ app/Http/Controllers/web/StudentController.php
@@ -1,10 +1,12 @@
 <?php
 namespace App\Http\Controllers\web;
 
+use App\helper\AccessRightsHelper;
 use App\Http\Controllers\Controller;
 
 use App\Center;
 use App\helper\ImageUploader;
+use App\Job;
 use App\QueryFilter\Name;
 use App\QueryFilter\Sort;
 use App\QueryFilter\SortElse;
@@ -27,36 +29,27 @@ use mysql_xdevapi\Session;
 class StudentController extends Controller
 {
 
+
     public function __construct()
     {
         $this->middleware('auth');
     }
 
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
 
-//        $this->authorize('viewAny',Student::class);
+        $this->authorize('view',Student::class);
 
         return view('students.all')->with('students',$this->getStudents());
     }
 
     public function getStudents(){
         $center = Center::findOrFail(Session('center_id'));
-//        dd($center);
         return Student::allStudents($center);
 
     }
 
-    /**
-     * Display a listing of the resource as table.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function showTable()
     {
         $center = Center::findOrFail(Session('center_id'));
@@ -64,34 +57,23 @@ class StudentController extends Controller
         return view('students.all-table')->with('students', $students);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
-//        dd(Session('center_id'));
-        //check if user has rights to view create_student_form
-        // $this->authorize('create',Student::class);
+        // authorization check
+        $this->authorize('create', Student::class);
+
         $student = new Student();
         return view('students.studentCreate',compact('student'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(Request $request)
     {
-        // todo : attach student to the center
         // check if user has rights to add a new student
+        $this->authorize('create', Student::class);
 
-//        $this->authorize('create',Student::class);
         $data = $this->validateRequest($request);
-
 
         // fetch center from session
         $center = Center::findOrFail(Session('center_id'));
@@ -112,44 +94,32 @@ class StudentController extends Controller
     }
 
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function show(Student $student)
     {
-//        $this->authorize('view',$student);
+        // check if student related to the auth center
+        $center = Center::findOrFail(Session('center_id'));
+        if (! $center->students->contains($student)){
+            return abort(404);
+        }
+
+        // authorization check
+        $this->authorize('view', Student::class);
         return view('students.show',compact('student'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit(Student $student)
     {
-
-//        $this->authorize('update',$student);
+        $this->authorize('update', $student);
         return view('students.studentEdit',compact('student'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, Student $student)
     {
 
-//        $this->authorize('update',$student);
-
-        // todo delete prev image from profiles dir
+        $this->authorize('update',$student);
 
         $data = $this->validateRequest($request);
 
@@ -170,7 +140,7 @@ class StudentController extends Controller
   function destroy(Student $student)
     {
         //policy
-//        $this->authorize('delete',$student);
+        $this->authorize('delete',$student);
         $student->delete();
         return response()->json('student deleted successfully', 200);
